pyproject.toml
[tool.poetry]
name = "vestapol"
version = "0.0.10"
description = "Python package that loads data from the web and deploys a corresponding external table definition, so that the data can be queried using standard SQL."
authors = ["Brian Waligorski <bwaligorski@inquirer.com>"]
license = "Apache License 2.0"
readme = "README.md"  # Markdown files are supported
repository = "https://github.com/phillymedia/vestapol"
homepage = "https://github.com/phillymedia/vestapol"

[tool.poetry.dependencies]
python = ">=3.8.12,<3.11"
requests = "^2.26"
google-cloud-storage = "^1.43"
google-cloud-bigquery = "^2.31"
pendulum = "^2.1"

[tool.poetry.dev-dependencies]
python-dotenv = "0.19.2"
pytest = "6.2.5"
tox = "3.25.0"
mypy = "^0.950"
types-requests = "^2.27.26"
flake8 = "^4.0.1"
black = "^22.3.0"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
